Exercise 1.3: added JavaScript
Added a for loop to the script to output the name and height of each Pokemon in the array.

Also, included an if conditional so that Pokemon larger than "3" would add "Wow! That's big!" next to the data.

Changed title of index.html to read "Pokemon Pokedex Pokedata" as a temporary, yet more descriptive, title.