Exercise 1.5 Part 2
Updated script.js to use and IIFE with keys to add new entry and print all Pokemon

Also adjusted height values to use hundredth of a decimal value.

Created typeof check using if to make sure new data is an object. Attempted to add new item that is a string to see if it will be added.

Added line to print getAll to console.log. String was not added to master list.

Commented out old code, will delete for next assignment.