Exercise 1.7: API added
Added fetch for pokemon API.

Index now lists 150 Pokemon.

Changed showDetails function to log Pokemon data in console when clicked.

** Also deleted some old code **

Added two functions to script for a LOADING message that appears and disappears when: site is loaded/Pokemon is clicked to show details.

However, the data loads so quickly that the LOADING message is impossible to read within the time it is visible. Regardless, the functions work.

Added polyfill scripts for fetch and promise functions to support users with older browsers