Adjusted Loading Message
Adjusted properties of Loading Message to be visible longer by using setTimeOut(). Also adjusted style properties to look better.

The LOADING message now scrolls with the page. However:

Tried mutltiple methods to center the message on the page, but it still is at the top left.

The styling of the page will change drastically and the message might not even be there by the time bootstrap is implemented on the page.